+++ projects/ppocr/ppocr/strategies/dcompass.cpp
@@ -1,16 +1,16 @@
 /*
 * Copyright (C) 2016 Wallix
-* 
+*
 * This library is free software; you can redistribute it and/or modify it under
 * the terms of the GNU Lesser General Public License as published by the Free
 * Software Foundation; either version 2.1 of the License, or (at your option)
 * any later version.
-* 
+*
 * This library is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
 * details.
-* 
+*
 * You should have received a copy of the GNU Lesser General Public License along
 * with this library; if not, write to the Free Software Foundation, Inc., 59
 * Temple Place, Suite 330, Boston, MA 02111-1307 USA
@@ -31,10 +31,10 @@ namespace ppocr { namespace strategies {
 static unsigned count_dcompass(Bounds const & bnd, Pixel const * p, Pixel const * ep, bool is_top)
 {
     unsigned d = 0;
-    size_t ih = 0;
-    size_t const wdiv2 = bnd.w()/2;
+    unsigned ih = 0;
+    unsigned const wdiv2 = bnd.w()/2;
     for (; p != ep; p += bnd.w(), ++ih) {
-        size_t x = wdiv2 - bnd.w() / (!is_top ? bnd.h() - ih : 1 + ih) / 2;
+        unsigned x = wdiv2 - bnd.w() / (!is_top ? bnd.h() - ih : 1 + ih) / 2;
         auto leftp = p + x;
         auto rightp = p + bnd.w() - x;
         for (; leftp != rightp; ++leftp) {
